收益问题修改

+++ WebRoot/js/dtjk/elevator.js
@@ -42,7 +42,9 @@
          }
 		//删除
 		  function del(id){
-    		  window.location.href="elevatorAction.do?method=delEntity&id="+id;
+			  if( confirm("确定删除这条电梯记录？")){
+    		 		 window.location.href="elevatorAction.do?method=delEntity&id="+id;
+    		  }
          }
 		  
 		 //下载  
@@ -75,12 +77,16 @@
 		  	//		批量删除
 $(".del").click(function() {
 	var ids="";
+	var len=0;
 	$(".wei").each(function() {
 			if($(this).children("i").hasClass("gou")) {
 				ids=ids+$(this).find("i").children().val()+",";
+				len++;
 			}
 		})
-	 window.location.href="elevatorAction.do?method=deleteEntity&ids="+ids;
+		 if(len>0 && confirm("确定删除这 "+len+" 条记录？")){
+	 		window.location.href="elevatorAction.do?method=deleteEntity&ids="+ids;
+	 	}
 })
